+++ src/FhirStore.CommonVersioned/Operations/OpSubscriptionHook.cs
@@ -31,8 +31,11 @@ public class OpSubscriptionHook : IFhirOperation
     /// <summary>Gets a value indicating whether this operation is a named query.</summary>
     public bool IsNamedQuery => false;
 
+    /// <summary>Gets a value indicating whether this operation affects the state of the store.</summary>
+    public bool AffectsState => true;
+
     /// <summary>Gets a value indicating whether we allow get.</summary>
-    public bool AllowGet => true;
+    public bool AllowGet => false;
 
     /// <summary>Gets a value indicating whether we allow post.</summary>
     public bool AllowPost => true;
@@ -143,6 +146,7 @@ public class OpSubscriptionHook : IFhirOperation
             Url = CanonicalByFhirVersion[fhirVersion],
             Status = Hl7.Fhir.Model.PublicationStatus.Draft,
             Kind = IsNamedQuery ? Hl7.Fhir.Model.OperationDefinition.OperationKind.Query : Hl7.Fhir.Model.OperationDefinition.OperationKind.Operation,
+            AffectsState = AffectsState,
             Code = OperationName.Substring(1),
             Resource = SupportedResources.CopyTargetsNullable(),
             System = AllowSystemLevel,