+++ src/FhirStore.R4/Operations/OpPasClaimSubmit.cs
@@ -36,6 +36,9 @@ public class OpPasClaimSubmit : IFhirOperation
     /// <summary>Gets a value indicating whether this operation is a named query.</summary>
     public bool IsNamedQuery => false;
 
+    /// <summary>Gets a value indicating whether this operation affects the state of the store.</summary>
+    public bool AffectsState => true;
+
     /// <summary>Gets a value indicating whether we allow get.</summary>
     public bool AllowGet => false;